Add zoom.translateExtent.
The translate extent, unlike the scale extent, isn’t a simple constraint on t_x
and t_y; instead, it is the extent of the viewable world. The zoom behavior can
then limit the transform such that the extent of the view (zoom.extent) is
always within the extent of the world.

This also replaces zoom.size with zoom.extent so that you can tell the zoom
behavior the extent of the viewport, rather than assuming the top-left corner is
always at [0, 0].

Fixes #25. Related d3/d3#1084 d3/d3#1299.