Don’t share state across elements.
The previous implementation assumed that there could only be one mouse or one
wheel active concurrently. While this feels like a reasonable assumption, it was
possible to break if the zoom behavior was applied to a parent and child element
simultaneously, since these would appear as concurrent gestures. Now the state
is isolated on the per-element gesture. Fixes #41.